requirements.txt
flask
flask-jwt-extended
werkzeug
pandas
hashlib
nltk
pyspellchecker
textblob
nltk